requirements.txt
Pillow==5.3.0
pytesseract==0.2.5
XlsxWriter==1.1.2